Fixed bugs. Added new features.
Fixed bug with immigration operator.
Added 2 new topologies and a way to select between them.
Added new configuration param: executions that allow to execute the algoritm n times.

+++ Util.cpp
@@ -42,7 +42,7 @@ vector<int> Util::random(int range, int size) {
 	thread_local std::mt19937 generator(rd());
 	std::uniform_int_distribution<int> dis(0, range - 1);
 
-	while (s.size() < size) {
+	while (s.size() < (unsigned int) size) {
 		int r = dis(generator);
 		s.insert(r);
 	}
